Fix issue472: make JsonStream a template.
